requirements.txt
streamlit==1.36.0

langchain==0.0.352
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87

sentence-transformers==2.7.0
chromadb==0.4.24
faiss-cpu==1.7.4

scrapling==0.4.6
beautifulsoup4==4.12.3
requests==2.31.0

google-genai==1.10.0

python-dotenv==1.0.1
tqdm==4.66.4
numpy==1.26.4